requirements.txt
# backend/requirements.txt
Flask
Flask-Cors
python-dotenv
google-generativeai
python-docx
PyMuPDF
reportlab
Pillow
requests
weasyprint
beautifulsoup4
jinja2
Flask
Flask-Cors
python-docx
pypdf
beautifulsoup4
jinja2
weasyprint
google-generativeai
python-dotenv